Add permission for timetable

+++ DUT/DUT.Web/Controllers/V1/TimetableController.cs
@@ -1,5 +1,6 @@
 ﻿using DUT.Application.Services.Interfaces;
 using DUT.Application.ViewModels.Timetable;
+using DUT.Constants;
 using Microsoft.AspNetCore.Mvc;
 
 namespace DUT.Web.Controllers.V1
@@ -7,17 +8,20 @@ namespace DUT.Web.Controllers.V1
     [ApiVersion("1.0")]
     public class TimetableController : ApiBaseController
     {
-        private readonly IAuthenticationService _authenticationService;
+        private readonly IPermissionService _permissionService;
         private readonly ITimetableService _timetableService;
-        public TimetableController(IAuthenticationService authenticationService, ITimetableService timetableService)
+        public TimetableController(IPermissionService permissionService, ITimetableService timetableService)
         {
-            _authenticationService = authenticationService;
+            _permissionService = permissionService;
             _timetableService = timetableService;
         }
 
         [HttpGet]
         public async Task<IActionResult> GetGroupTimetable(int groupId, DateTime? from, DateTime? to)
         {
+            if (!_permissionService.HasPermission(PermissionClaims.Timetable, Permissions.CanView))
+                return JsonForbiddenResult();
+
             if (from == null)
                 from = DateTime.Today;
             if (to == null)
@@ -33,18 +37,27 @@ namespace DUT.Web.Controllers.V1
         [HttpPost]
         public async Task<IActionResult> CreateTimetable([FromBody] TimetableCreateModel timetable)
         {
+            if (!_permissionService.HasPermission(PermissionClaims.Timetable, Permissions.CanCreate))
+                return JsonForbiddenResult();
+
             return JsonResult(await _timetableService.CreateTimetableAsync(timetable));
         }
 
         [HttpPut]
         public async Task<IActionResult> UpdateTimetable([FromBody] TimetableCreateModel timetable)
         {
+            if (!_permissionService.HasPermission(PermissionClaims.Timetable, Permissions.CanEdit))
+                return JsonForbiddenResult();
+
             return JsonResult(await _timetableService.UpdateTimetableAsync(timetable));
         }
 
         [HttpDelete]
         public async Task<IActionResult> DeleteTimetable(long[] ids)
         {
+            if (!_permissionService.HasPermission(PermissionClaims.Timetable, Permissions.CanRemove))
+                return JsonForbiddenResult();
+
             return JsonResult(await _timetableService.RemoveTimetableAsync(ids));
         }
     }
